苍穹外卖圆满收工，可喜可贺，可喜可贺

+++ sky-server/src/main/java/com/sky/mapper/UserMapper.java
@@ -26,5 +26,5 @@ public interface UserMapper {
 	@Select("select * from user where id=#{userId}")
 	User getById(Long userId);
 
-	Double sumByMap(Map map);
+	Integer sumByMap(Map map);
 }
